backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<1.0
django-filter>=24.0,<25.0
supabase>=2.0,<3.0
PyJWT>=2.8,<3.0
psycopg2-binary>=2.9,<3.0
gunicorn>=22.0,<23.0
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
httpx>=0.27,<1.0
flake8>=7.0,<8.0
